requirements.txt
flask==2.3.3
werkzeug==2.3.7
firebase-admin==6.2.0
python-dotenv==1.0.0
requests==2.29.0
flask-login==0.6.3
pyrebase4==4.7.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
requests-toolbelt==0.10.1
urllib3<2.0.0
cryptography==41.0.7
setuptools==69.2.0
wheel==0.42.0
python-jwt==4.0.0
pycryptodome==3.19.1
google-cloud-storage==3.0.0
google-auth==2.28.1
google-api-python-client==2.108.0
oauth2client==4.1.3 